# Conflicts:
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/.vs/Projet_Huchon_Salemi_3I/DesignTimeBuild/.dtbcache.v2
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/.vs/Projet_Huchon_Salemi_3I/v16/.suo
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/View/MembreVIEW.xaml.cs
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll.locked
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.pdb
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.g.resources
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.pdb
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I_MarkupCompile.i.cache
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I_MarkupCompile.i.lref
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/View/HomeVIEW.baml
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/View/HomeVIEW.g.cs
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/View/MembreVIEW.baml
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/View/MembreVIEW.g.cs

+++ Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/metier/Balade.cs
@@ -9,7 +9,7 @@ namespace Projet_Huchon_Salemi_3I.metier
     class Balade
     {
         public decimal id_vehicule;
-        private int num;
+        private decimal num;
         private string lieuDepart;
         private DateTime dateDepart;
         private decimal forfait;
@@ -21,7 +21,7 @@ namespace Projet_Huchon_Salemi_3I.metier
         {
         }
 
-        public Balade(int num, string lieuDepart, DateTime dateDepart, decimal forfait)
+        public Balade(decimal num, string lieuDepart, DateTime dateDepart, decimal forfait)
         {
             this.num = num;
             this.lieuDepart = lieuDepart;
@@ -29,7 +29,7 @@ namespace Projet_Huchon_Salemi_3I.metier
             this.forfait = forfait;
         }
 
-        public Balade(int num, string lieuDepart, DateTime dateDepart, decimal forfait, List<Vehicule> listeVehicule, List<Inscription> listeInscription, decimal id_calendrier)
+        public Balade(decimal num, string lieuDepart, DateTime dateDepart, decimal forfait, List<Vehicule> listeVehicule, List<Inscription> listeInscription, decimal id_calendrier)
         {
             this.num = num;
             this.lieuDepart = lieuDepart;
@@ -40,7 +40,7 @@ namespace Projet_Huchon_Salemi_3I.metier
             this.CalendrierBalade = id_calendrier;
         }
 
-        public Balade(int num)
+        public Balade(decimal num)
         {
             this.num = num;
         }
@@ -53,7 +53,7 @@ namespace Projet_Huchon_Salemi_3I.metier
         }
 
 
-        public int Num { get => num; set => num = value; }
+        public decimal Num { get => num; set => num = value; }
         public string LieuDepart { get => lieuDepart; set => lieuDepart = value; }
         public DateTime DateDepart { get => dateDepart; set => dateDepart = value; }
         public decimal Forfait { get => forfait; set => forfait = value; }
